Rename configuration name
