--- src/types/collapsedForm.ts
@@ -1,13 +1,0 @@
-/**
- * Configuration for collapsed form
- */
-export type CollapsedFormConfig = {
-  /**
-   * Title for collapsed form
-   */
-  title?: string;
-  /**
-   * Description for collapsed form
-   */
-  description?: string;
-};
