+++ src/types/widget.ts
@@ -1,9 +1,13 @@
-import { NotionConfig } from '../service/notion/config';
-import { FormConfig } from './form';
-import { CollapsedFormConfig } from './collapsedForm';
-
-export type Configuration = {
-  form: FormConfig;
-  collapsedForm: CollapsedFormConfig;
-  notion: NotionConfig;
+/**
+ * Configuration for collapsed form
+ */
+export type WidgetConfig = {
+  /**
+   * Title for collapsed form
+   */
+  title?: string;
+  /**
+   * Description for collapsed form
+   */
+  description?: string;
 };
